Adding additional documentation to the ListT and ListTElement classes; everything seems to be working except the AND operator when the expressions involved are not in brackets.
Example:
int a;
a = 10;
print(a);

if((a < 100) & (a = 10))
  print(1)
else
  print(0);

displays what is expected:
10
1

However when the brackets in the if-condition are removed following is
displayed:
10
0

Swapping the expressions didn't change this behaviour. Everytime when no
brackets are used this leads to a false evaluation of the condition
because of the AND operator. The question is - is this how it's supposed
to work in the given language?